fix: (router) refactor cmd/main to expose router instance
This enables consumers to own the router lifecycle without needing access to the internals of the router library.
